Improve Hotloading Support
- Changes hotloader to be a lib rather than a bin so that client
libraries can list it as a dependency and compile it automatically.
- Removes the hard-coded path to the game dll from loader in favor of
having the client lib pass it in when starting the loader.
- renames polygon_rs crate to polygon in order to avoid breakage with
hotloading. It probably should just be polygon, anyway, so not a big
deal.
- Adds a null module to the stopwatch crate so that client code can
disable usage of stopwatch without removing all calls to its code.
- Adds "timing" and "hotloading" features to gunship to allow client
code to enable and disable timing critical sections and hotloading
respectively.
- Fixes hotloading by using a modified version of the type name when the
hotloading feature is enabled. When hotloading is disabled the TypeId is
used instead because it is cheaper at runtime.