refactor(gunship::resource): split resource parsing into separe modules
Previously the resource manager (in the gunship::resource) directly held
the code for parsing COLLADA files and shaders. Move the shader code into
gunship::resource::shader and the COLLADA code into
gunship::resource::collada. Also begin work on improving the COLLADA
parsing code to not be so hard-coded and broken.