refactor(Camera): create camera component
Have `CameraManager` use its own `CameraType` rather than polygon's
`Camera` type directly so that it doesn't expose the position and
orientation through the camera which could lead to counterintuitive bugs
in game code.

BREAKING CHANGE: Hotloading is broken as a result of this change until we
have a better system for doing compilation-stable type ids.