refactor(alarm): don't require mutable borrow on alarm manager
Change alarm manager to be able to handle adding and removing alarms with
a shared borrow. Also fix hotloading support for alarm manager with new
callback utilities.

BREAKING CHANGE: `AlarmManager` now requires that an alarm callback be
registered with `register_callback()` before calling `assign()` or it will
panic. This will not cause a compile error, but will cause a panic.