Add Support For Hiding/Capturing The Cursor
Add some preliminary utlities for hiding the cursor while it's over the
game window and capturing it so that it does not leave the window. This
support is still not ideal and the interface is not final but it gives
a basic idea of what the final design will look like.