Render Reform (#70)
closes #65
closes #64
closes #61
closes #56
closes #49

This was a major reform pass that makes huge changes to the rendering system. Polygon was almost completely rewritten, including writing new custom OpenGL bindings and a wrapper library to provide a safe abstraction over raw OpenGL procs. This basically *is* the Basic Rendering milestone. This PR is way too big but it's too late to try to break it up now.

Also, this PR completely breaks the engine. Basically nothing works now. I didn't bother fixing up everything because we've got even more major reform planned for the engine proper so there's no point in fixing code that's about to be thrown out.